# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Ontrek backend",
        "title": "Ontrek API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/groups/": {
            "get": {
                "description": "Get all groups for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Get groups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token for user authentication",
                        "name": "Bearer",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of groups",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/utils.GroupDoc"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error fetching files",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new group for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Create a new group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token for user authentication",
                        "name": "Bearer",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Group information. Fields: description (required, max 64 characters), file_id (optional, must be a valid file ID)",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.GroupInfoCreation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Group created successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.GroupMemberCreation"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/groups/{id}/members/{user}": {
            "put": {
                "description": "Allows a user to join a group by providing their group ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Join a group using group ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token for user authentication",
                        "name": "Bearer",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User successfully added to the group",
                        "schema": {
                            "$ref": "#/definitions/utils.GroupMember"
                        }
                    },
                    "400": {
                        "description": "Group ID is required",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Only group leader can add members",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Group not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "User already joined the group",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "An error occurred"
                }
            }
        },
...
        "utils.GpxInfoEssential": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "MonteBianco"
                }
            }
        },
        "utils.GroupDoc": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-05-11T08:00:00Z"
                },
                "created_by": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "description": {
                    "type": "string",
                    "example": "Morning hike with friends"
                },
                "group_id": {
                    "type": "integer",
                    "example": 1
                },
                "members_number": {
                    "type": "integer",
                    "example": 5
                },
                "track": {
                    "$ref": "#/definitions/utils.GpxInfoEssential"
                }
            }
        },
        "utils.GroupInfoCreation": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Morning hike with friends"
                },
                "file_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "utils.GroupMember": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string",
                    "example": "#e6194b"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "username": {
                    "type": "string",
                    "example": "John Doe"
                }
            }
        },
        "utils.GroupMemberCreation": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string",
                    "example": "#e6194b"
                },
                "group_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "username": {
                    "type": "string",
                    "example": "John Doe"
                }
            }
        },
...
    }
}